// apps/mobile/src/navigation/types.ts
export type RootStackParamList = {
  Login: undefined;
  Register: undefined;
  MainTabs: undefined;
  CreateSubmission: undefined;
};

export type MainTabParamList = {
  Home: undefined;
  Themes: undefined;
  Submissions: undefined;
  Hotspots: undefined;
  Rankings: undefined;
  Insights: undefined;
  Profile: undefined;
};